Add not found group and user validations

+++ DocumentsStore/DocumentsStore/UseCases/Users/RemoveUserFromGroup.cs
@@ -7,16 +7,35 @@ namespace DocumentsStore.UseCases.Users;
 public class RemoveUserFromGroup : IRemoveUserFromGroup
 {
     private readonly IGroupUsersRepository _groupUsersRepository;
+    private readonly IGroupsRepository _groupsRepository;
+    private readonly IUsersRepository _usersRepository;
 
-    public RemoveUserFromGroup(IGroupUsersRepository groupUsersRepository)
+    public RemoveUserFromGroup(IGroupUsersRepository groupUsersRepository, IGroupsRepository groupsRepository, IUsersRepository usersRepository)
     {
         _groupUsersRepository = groupUsersRepository;
+        _groupsRepository = groupsRepository;
+        _usersRepository = usersRepository;
     }
     
     public async Task<UseCaseResult<User>> ExecuteAsync(int userId, int groupId, CancellationToken cancellationToken)
     {
-        var result = await _groupUsersRepository.RemoveUserFromGroup(userId, groupId, cancellationToken);
+        var user = _usersRepository.GetByIdAsync(userId, cancellationToken);
+        var group = _groupsRepository.GetByIdAsync(groupId, cancellationToken);
 
-        return UseCaseResult<User>.Success(result);
+        await Task.WhenAll(user, group);
+        
+        if (await user is null)
+        {
+            return UseCaseResult<User>.NotFound("User not found");
+        }
+
+        if (await group is null)
+        {
+            return UseCaseResult<User>.NotFound("Group not found");
+        }
+        
+        await _groupUsersRepository.RemoveUserFromGroup(userId, groupId, cancellationToken);
+
+        return UseCaseResult<User>.Success((await user)!);
     }
 }
